port seriesHashmap and stripeSeries from TSDB
stripeSeries will be used by the wal.Storage to determine when a series
no longer receives updates so it can be deleted from subsequent
checkpoints.

This commit does most of the work of deleting churned series from the
checkpoint, but doesn't actually do it - until WAL replaying happens on
startup, it's unsafe to actually throw away the series. A subsequent
commit for replaying the WAL will be coupled with enabling the
deletions.