only delete series from stripeSeries after a second gc
Removing series immediately when the timestamp is based on last remote
write time doesn't make sense. There's always going to be some period
where the remote write timestamp is after the most recent scrape
timestamp, causing everything to be marked for deletion.

If these series are deleted immediately, once the next scrape runs, they
will be added as new series and written to the WAL as such, even if the
last record has yet to be removed.

This solution isn't perfect but takes care of the normal flow. See the
added TODO for how to implement this better.