add prometheus' interner
this commit introduces the Prometheus interner found in its remote
storage code. The interner is added as a separate package and is used to
intern labels when replaying from the WAL.

this reduces memory usage of replaying from the WAL dramatically, but
doesn't remove it all together, as labels are still being interned
twice: once by the WAL and once more by the remote_write Queue Manager.