Set the scrape manager for the remote storage
Typically (within Prometheus and here in the agent) we initialise the remote storage before we initialise the scrape manager. Since sending metadata relies on the scrape manager, this commit introduces the use of a `ReadyScrapeManager`.

The `ReadyScrapeManager` allows us to set the scrape manager eventually while allowing the remote storage to access it once it is set.

Signed-off-by: gotjosh <josue@grafana.com>