pkg/prometheus: fix host_filter not matching `role: node` k8s SD jobs
Previously, the `host_filter` code was looking for a label called
__meta_kubernetes_pod_node_name. This label is only available when using
the `role: pod` matcher within the kubernetes_sd_configs config block.

When using `role: node`, which is common for the cadvisor and kubelet
jobs, the metalabel for the node name is instead called
`__meta_kubernetes_node_name`. host_filter wasn't checking for this
label and so any `role: node` job was getting filtered out.