add scraping service node resharding
This commit adds support for resharding a node in the scraping service
agent cluster. This is implemented using gRPC servers that expose a
Reshard RPC method. While the gRPC servers are tiny right now, it is
anticipated that more gRPC methods will be added in the future.

The scraping service node lifecycle is now complete with this commit.
When the agent starts, it does the following:

1. Join the Cortex ring
2. Connect to all nodes in ring (except self)
3. Tell nodes to reshard

When the agent is running, it reshards itself on an interval. This is
done so config drift against the KV store is caught at runtime so the
agents will eventually correct themselves in case of a bug.

When the agent shuts down, it does the following:

1. Mark itself as leaving in the Cortex ring.
2. Connect to all nodes in ring (except self)
3. Tell nodes to reshard.
4. Leave the Cortex ring

The resharding process is accomplished as follows:

1. Retrieve all currently running configs.
2. Retrieve all configs currently registered in the KV store.
3. For each config in the config store, check to see if that config
   needs to be removed, added, or changed within the local agent.
4. Remove any remaining configs that are currently registered but
   not present within the config store.