pkg/prometheus/ha: store and hash secrets properly (#63)
The Bearer Token and Basic Auth Password for remote_write configs were
not being stored or hashed properly; those values are of a Secret type
that implements a custom yaml.Marshaler and removes the underlying
secret value.

So when configs were being stored in the KV store, the secrets were
being removed and replaced with the text "<secret>". This caused
agents who loaded the config to run an instance to use the wrong
password.

Hashing a config had the same problem; since the hash only marshaled a
config to YAML, secrets were not being calculated properly with the
hash.

This commit fixes both problems; secrets are stored separately now in a
raw string type and hashing adds the values for secrets before summing
the input.

Fixes #62.