simplify instance creation and lifecycle
As part of #83, the instance package will have to be utilized in a new
way for running integrations (i.e., running an instance that does not
have a discovery component).

To prepare for this change, the instance package was in need of a
refactoring to simplify how the instance objects are used. Now, instead
of an instance immediately starting when it is created, it must be
started explicitly through the Run function. Run is a blocking operation
and will only exit on error or when the context is canceled.  Changing
to this allowed for cleaning up a lot of code callers required
to watch the lifecycle of the Instance.

Note that the code for unregistering WAL metrics is no longer needed as
the WAL is now created when Run is called.

Running an instance with a discovery manager is still required as of
this commit; future work is still needed to make the individual
components of the Instance optional.