use remote_write and scrape_config validations from Prometheus
Prometheus was doing some validations on the remote_write and
scrape_config sections per config that we weren't doing yet. To make
this a little easier, ApplyDefaults and Validate have been merged.

The scraping service PutConfiguration endpoint will now apply the
defaults to the config before passing it through to the KV store;
invalid configs will be prevented from reaching the KV store and errors
will be returned to the client.

When migrating to this change, if any invalid configs exist in the KV
store, the Agent will refuse to start them and log an error. The configs
in the KV store should be updated to fix the invalid settings before the
Agent will be able to run it.

Fixes #80.