sharding config manager should only list its own configs
The introduction of the integrations system will present the opportunity
for the prom.Agents' ConfigManager to be responsible for a set of
configurations from the scraping service KV store and a set of
configurations from the integrations.

Since the scraping service mode uses ListConfigs to determine if a
running config has been deleted from the KV store, we don't want any
integrations configs to be a part of that deleted set. To work around
this, the sharding config manager should only list configs that have
been applied through it, effectively giving it a namespace from the
overall set of configs managed by the underlying ConfigManager.