allow InstanceManager.ApplyConfig to fail
Previously, InstanceManager.ApplyConfig could not fail, but the instance
config would be ignored and log an error message if the config was
invalid. This process has now been separated and validation of the
config will be done before the goroutine is launched, allowing
ApplyConfig to return an error if the config was invalid.