Add integrations system (#93)
* sharding config manager should only list its own configs

The introduction of the integrations system will present the opportunity
for the prom.Agents' ConfigManager to be responsible for a set of
configurations from the scraping service KV store and a set of
configurations from the integrations.

Since the scraping service mode uses ListConfigs to determine if a
running config has been deleted from the KV store, we don't want any
integrations configs to be a part of that deleted set. To work around
this, the sharding config manager should only list configs that have
been applied through it, effectively giving it a namespace from the
overall set of configs managed by the underlying ConfigManager.

* finish working prototype

* add example config for running agent with integrations

* rename ConfigManager to InstanceManager

InstanceManager is a clearer name for what ConfigManager does: manage
running instances through their configs.

* allow InstanceManager.ApplyConfig to fail

Previously, InstanceManager.ApplyConfig could not fail, but the instance
config would be ignored and log an error message if the config was
invalid. This process has now been separated and validation of the
config will be done before the goroutine is launched, allowing
ApplyConfig to return an error if the config was invalid.

* change integration label values to be less ugly

integration/agent is better than __integration_agent :)

* finish main implementation of integrations system

- Add agent_hostname label by default (configurable) when using
  integrations

- Add a set of extra labels to add when scraping integrations

- Add scrape timeout and interval to a common set of integrations
  configs.

- Integrations now return a subset of scrape configs that will be
  used to generate more useful job names.

* add tests, fix problems found in testing

* shutdown integrations when server stops

* add docs for integrations

* add .idea to gitignore

* fix lint errors

* address review feedback