Fix zero default scrape interval by initializing global config stanza with defaults (fixes #106) (#107)
* Log message language fix

* Simplify error return

* config: Simplify first ApplyDefaults loop

This way, we're not accessing the slice via both range variable
and index.

* config: Remove redundant second ApplyDefaults loop

As far as I can tell, this is entirely redundant with the
first loop, which already calls ApplyDefaults on each of
the sub-configs.

* config: Test that scrape interval defaults to non-zero

* config: Apply global defaults on unmarshal, fixing zero scrape_interval

This adds a dedicated with-defaults UnmarshalYAML, modeled on
the one in prom/instance/instance.go. It ensures that Global
is initialized with the prometheus DefaultGlobalConfig, which
is subsequently used to provide defaults for ScrapeInterval.