allow dynamic updating of instance scrape_configs and remote_write
This commit introduces an Update method against the Instance that allows
for applying new changes found on remote_write and scrape_configs
sections for a particular instance. If any other field changes, Update
will return ErrInvalidUpdate, which will cause the InstanceManager to
do a normal "restart" of the instance.

Future commits may loosen the restriction on what is able to be
dynamically updated.

Some code cleanup can be found as part of this commit: fakes that have
been called mocks have been updated to use their proper name.