move InstanceManager to instance package
This commit migrates InstanceManager to instance.BasicManager. This is
to reduce duplication as we add more implementations of the
instance.Manager interface (including removing ha.InstanceManager, which
was a duplicate of that interface).

While creating this commit, I noticed that there was a leaked locked
mutex when DeleteConfig fails. That has been fixed with this commit.

Related to the work for #141.
Depends on #145.