pkg/loki: return empty loki when no configuration is present
If Loki isn't configured, then promtail.New will fail since it expects
at least one client config to be present. However, this means that no
Loki instance is created and calling stop on it is a nil reference,
throwing a panic.

This commit returns an empty Loki when Promtail isn't configured, and
explicitly exits if loki.New returns an error.

Fixes #153.