fix panic when starting scraping service (#177)
Updating the Cortex vendor for importing Loki broke the scraping
service. Two KV clients get initialized: one directly for storing
configs and one indirectly by creating a Cortex ring client. The Cortex
ring client wraps the registerer passed to the KV client with a call to
kv.RegistererWithKVName. Not doing this leads to an inconsistency on
constant labels added to the KV metrics and causes a panic from the
Prometheus client_golang package.