override remote_name when grouping configs (#175)
* override remote_name when grouping configs

In the instance sharing mode, remote_write configs with custom names
(remote_names) were being copied from the first element of the group.
This causes the remote_name to change the first element is removed,
which is an unexpected behavior. This particularly affects scraping
service users, where the stored configs already have defaults applied,
including the generated remote_name.

remote_names are now forced to be the combination of the fix six
characters of the group name and remote_write config hash separated by a
hypen. A future enhancement may wish to consider non-autogenerated names
when grouping remote_write configs, but that change is out of scope
here.

* Update CHANGELOG.md

Co-authored-by: Karsten Jeschkies <k@jeschkies.xyz>