fix Prometheus component update order when updating an Instance config
Prometheus applies component updates in a specific order. In particular,
the scrape manager needs to be applied before the discovery manager. If
this order is not kept, there's a chance the discovery manager will send
a target for a new job to the scrape manager before the scrape manager
is aware of the job. If this happens, the scrape manager drops all
targets associated with that job and will not retry.