Allow overriding instance label from replace_instance_label option
This commit changes the mechanism for which the replace_instance_label
config option is applied; rather than being applied through
write_relabel_configs, it is now applied as the first relabel_config
when relabeling targets.

This allows users to replace the instance label with their own value,
and even use the existing overwritten instance label for creating the
final result.

Closes #220