add min_wal_time and max_wal_time to bound WAL data lifetime (#253)
* support ensuring data lives in the WAL for a configurable amount of time

It's suspected that immediately deleting data from the WAL as soon as it
is remote_written is error prone, with one potential scenario described
in #248.

This PR adds a new flag to control how long data must live for in the
WAL before it can be considered ready for deletion. The default setting
is 5m, which should be ample time for remote_write to give up on a
failed batch. Users are free to set this value to anything, including
0s, if they want to retain the old behavior, or 24h, if they want a huge
replay buffer for their WAL.

Closes #138.

* explain logic behind subtracing MinWALTime

* max_wal_time

* fix failing test