scraping service: improve sharding mechanism
Sharding will now be performed based on the name of the config file
instead of the contents of the config file. This guarantees that in a
static ring, the same config name will always hash to the same Agent.
This improves guarantees of the hashing, but also ensures that changing
a config will not cause it to move to another Agent and cause gaps in
metrics as the scrape responsibility moves.