metrics: read global config when reloading instances
This PR fixes an issue where the global settings were not being read
by instances when reloading. This was because the global settings were
passed to the constructor of instances, but never updated when applying
the new instance settings.

This PR fixes that problem by adding the global config to be a part of
the instance config struct. It cannot be marshaled into the instance,
and must be set manually by the code before applying the changes.

There are two code paths to consider here:

1. The scraping service will re-read the global config on the next
   reshard tick.
2. Non-scraping service will apply the global config as soon as the new
   config is read in.

This also changes the group manager to ensure the global settings are
kept when creating a group.