TestEnqueueRequestForSelector: only run test if -online flag is set
This test does download and invoke kubebuilder-tools, to make sure there
is no extra steps needed just to run this one test.
(https://github.com/grafana/agent/issues/731#issuecomment-878255964).

However, distros usually build and run tests in a sandbox with network
access disabled, causing this test to fail.

Introduce a `online` flag, defaulting to false, and skip this test as
long as `-online` is not passed.

CI, which invokes `make test` deliberately sets this to true, to have
coverage for that test provided by CI.

Fixes #731.