Move port to load_balancing config as receiver_port
When extracting load balancing from tail sampling,
only the the load_balancing block was moved into its own block,
but the receiver's port was left in tail_sampling.

This means that the receiver port for load_balancing could not
be configured without tail_sampling.

Now its moved to load_balancing as receiver_port.