Update for `## Deploy GrafanaAgent` clusterrole (#923)
Update `## Deploy GrafanaAgent` section to include additional kubelet metrics.  Reason:  When checking the metrics sent to Grafana Cloud, we noticed that some of kubelet metrics (cpu/memory usage) was missing.  On the agent, we are using the backwards compatibility with the Prometheus Operator CRDs ServiceMonitor and PodMonitor , so that metric instance discovery is exactly as it is with our hosted prometheus.