fork the cortex ring/kv package for use with the configstore
The cortex ring/kv package is focused heavily on a hash ring, making it
increasingly inappropriate to use as a KV store for the configstore,
which expects a ton of traffic. This initial PR does a basic copy
of the cortex package into configstore/kv.

Initial changes from the Cortex package:

- Memberlist is removed
- cortex_ metrics have been reprefixed with agent_