[dev.operator-integrations-support] Generate agent config for an integrations deployment (#1012)
* pkg/operator: generate agent config for an integrations deployment

Note that the Type field is unused here. Type determines whether a
DaemonSet or Deployment will be made, and doesn't change the generated
agent config.

* use consistent pluralization for "integration"

Note that the Type field is unused here. Type determines whether a
DaemonSet or Deployment will be made, and doesn't change the generated
agent config.