integrations: Add App Observability integration
NOTE: This integration is implemented using the v2 Integration interface

This commit is the first draft of integrating the [App Observability
Telemetry Receiver][1] into the agent. For now the following components
are supported:

Logs
====

The exporter forwards the logs received onto the HTTP endpoint into the
logs subsystem.

Exceptions
==========

Same as logs, but with fixed labels.

Metrics
=======

The integration is capable of reporting both some predetermined
performance metrics (web vitals in this iteration) but also some user
defined ones using a specific schema in the configuration file.
Internally, a Prometheus Summary type is used to keep track of the
incoming values and is also using the self scrape system to report the
metrics.

The exporter itself is basically a modular HTTP endpoint that uses
different exporters to forward received data into the appropriate
subsystems of the agent.

[1]: https://github.com/grafana/grafana-frontend-telemetry-receiver