operator: add CRDs for integrations
This commit introduces CRDs for integrations:

* MetricsIntegration is a child of GrafanaAgent and is used to specify a
  running metrics-based integration.

* IntegrationMonitor is a child of MetricsInstance and is used to
  specify which MetricsIntegrations should be scraped for metrics.

The CRDs have been added to the resource hierarchy and will
automatically be discovered when reconciling a GrafanaAgent resource.
While the resources are discovered, this commit does *not* run
integrations or collect metrics from them.

As a part of this change, the old config.Deployment type has been moved
to pkg/operator/apis/monitoring/v1alpha1. This allows it to take
advantage of the DeepCopy function generation, a method which was
previously manually maintained.