operator/hierarchy: stop using field selector when listing Secrets & ConfigMaps (#1340)
The initial implementation of hierarchy.KeySelector injected a
FieldSelector when listing Secrets and ConfigMaps to immediately return
the single object being queried for.

This causes a problem with the client generated by the
controller-runtime framework, where the client is wrapped in a cache and
field indexer (where only the namespace is indexed by default).

This commit avoids using the field selector and the index lookup. The
resulting behavior aligns more closely with discovering other resources
in the hierarchy (i.e., ServiceMonitors), where the List call is also
insufficient and needs post-processing via Matches to find the final
list of resources.

Given the controller-runtime client uses an informer for reads, all
relevant Secrets and ConfigMaps are already in-memory anyway, and using
the index for a faster List is a bit of an over-optimization at the
moment.