integrations-next: allow all integrations to be unmarshaled <name> or <name>_configs (#1463)
This relaxes the restrictions on how integrations are unmarshaled so
that any integration, regardless of type, can be unmarshaled as an JSON
object or an array.

This makes configurations like this technically valid:

integrations:
  node_exporter_configs:
  - {}

Such a configuration would not be recommended to users, but as there is
only one defined instance of node_exporter, it still satisfies the
singleton requirements. It is unlikely that we would document this is
possible.

Introducing this change makes it *much* easier for the operator to
generate configs for integrations, as we no longer have to switch
between the _configs and non-_configs key based on the integration being
generated.