pkg/server: replace pkg/util/server with pkg/server
pkg/server is a new top-level subsystem which exposes an HTTP and gRPC
server. It was forked off of weaveworks/common/server with direct
support for reloading settings we wish to keep as dynamic.

The following YAML fields are intended to be dynamically updatable:

* server.log_level
* server.log_format
* server.http_tls_config
* server.grpc_tls_config

Other fields are intended to be static and should only be set via
command line flag. Their YAML counterparts are now DEPRECATED and will
be removed in v0.26.0. As of this commit, reloading the config file will
fail if a non-dynamic field has changed in the server block.

Closes grafana/agent#1336.

Flag changes
------------

Many flag names corresponding to the server have changed.

The following flags are NEW:

* `-server.http.enable-tls`
* `-server.grpc.enable-tls`
* `-server.http.address`
* `-server.grpc.address`

The `-server.*-enable-tls` flags must now be passed to enable TLS for
HTTP and gRPC respectively. This is a change from the previous behavior
of dynamically enabling TLS when a certificate pair is provided.

`-server.http.address` and `-server.grpc.address` are intended to be
`host:port` addresses on which to listen for HTTP and gRPC traffic. They
replace the `-server.*-listen-address` and `-server.*-listen-port` flags.

The following flags have been RENAMED:

* `-server.log.source-ips.enabled` (renamed from `-server.log-source-ips-enabled`)
* `-server.log.source-ips.header` (renamed from `-server.log-source-ips-header`)
* `-server.log.source-ips.regex` (renamed from `-server.log-source-ips-regex`)
* `-server.http.network` (renamed from `-server.http-listen-network`)
* `-server.http.conn-limit` (renamed from `-server.http-conn-limit`)
* `-server.http.read-timeout` (renamed from `-server.http-read-timeout`)
* `-server.http.write-timeout` (renamed from `-server.http-write-timeout`)
* `-server.http.idle-timeout` (renamed from `-server.http-idle-timeout`)
* `-server.grpc.network` (renamed from `-server.grpc-listen-network`)
* `-server.grpc.conn-limit` (renamed from `-server.grpc-conn-limit`)
* `-server.grpc.max-recv-msg-size-bytes` (renamed from `-server.grpc-max-recv-msg-size-bytes`)
* `-server.grpc.max-send-msg-size-bytes` (renamed from `-server.grpc-max-send-msg-size-bytes`)
* `-server.grpc.max-concurrent-streams` (renamed from `-server.grpc-max-concurrent-streams`)

The following flags have been REMOVED:

* `-log.level` (use `server.log_level` in the YAML instead)
* `-log.format` (use `server.log_format` in the YAML instead)
* `-server.http-tls-cert-path` (use `server.http_tls_config` in the YAML instead)
* `-server.http-tls-key-path` (use `server.http_tls_config` in the YAML instead)
* `-server.http-tls-client-auth` (use `server.http_tls_config` in the YAML instead)
* `-server.http-tls-ca-path` (use `server.http_tls_config` in the YAML instead)
* `-server.grpc-tls-cert-path` (use `server.grpc_tls_config` in the YAML instead)
* `-server.grpc-tls-key-path` (use `server.grpc_tls_config` in the YAML instead)
* `-server.grpc-tls-client-auth` (use `server.grpc_tls_config` in the YAML instead)
* `-server.grpc-tls-ca-path` (use `server.grpc_tls_config` in the YAML instead)
* `-server.http-listen-address` (use the `-server.http.address` flag instead)
* `-server.http-listen-port` (use the `-server.http.address` flag instead)
* `-server.grpc-listen-address` (use the `-server.grpc.address` flag instead)
* `-server.grpc-listen-port` (use the `-server.grpc.address` flag instead)
* `-server.path-prefix` (usage of this flag was never supported and would cause issues)

`-server.path-prefix` and the equivalent YAML field `server.path_prefix`
have been removed. Setting this option would lead to undefined behavior
for various parts of the agent. Being able to prepend an API prefix may
return in the future if it is requested.

Flags not listed here are unchanged.

Other changes
-------------

All jsonnet and config file examples no longer use the deprecated field
names.