pkg/server: allow in-memory connections
This commit extends pkg/server with an in-memory net.Listener for HTTP
and gRPC. There is then a dialer function which can perform an in-memory
Dial with a specific connection address. The connection addresses used
for in-memory HTTP and gRPC can be overridden via flags if they collide
with a real DNS name.