pkg/operator: refactor pod template generation into its own function
Extract the generation of pod templates out into its own function: the
implementation of generating the DaemonSetSpec (for logs) and
StatefulSetSpec (for metrics) was incredibly similar, and this will be
doubly true for integrations. A shared function for generating the
template will help remove duplication and avoid bugs where template
generation becomes out of sync for the various generated resources.