pkg/config: use -config.file.type flag to derive file type
Introduce a -config.file.type which acts as a hint for the contents of
the config file being loaded. This then replaces the
-config.dynamic-config-path flag with -config.file.type=dynamic.

Having a dedicated flag for file should be useful for avoiding having
too many flags which load a config file and allows us to change the
default type in the future (i.e., default to `-config.file.type=flow`).