pkg/flow: cache cty.Values for components as they change
This cache introduces caching of the gocty values for the input and
output of a component. This avoids allocations where the input and
output is constantly converted to gocty despite the component not
changing. This is primarily useful when a component refers to multiple
nodes in its input, but only one of those nodes changes.

(This change also renames nametable to execContext as the latter is more
descriptive to what its purpose is.)