flow: initial commit of hcltypes package with Secret type
pkg/flow/hcltypes contains common types that components may wish to use.
The hcltypes.Secret type is used for strings holding sensitive values.
Under the hood, this is a capsule type: any string can be converted to
hcltypes.Secret, but the Secret may not be converted back into a string.