flow: initial commit for Flow controller
The Flow controller is a scheduler which runs and configures
"components" based on HCL files.

The set of running components can reference each other using HCL
expressions, forming a DAG.

The Flow controller will update components through the lifetime of the
process whenever a compnoent changes.

Components
----------

Components are units of business logic that have a set of input
arugments and optional exported fields. The inputs to a component
are decoded from HCL, while the exports of a component are encoded to
HCL.

A component may emit new values for its exports across the component's
lifetime. Other components which reference those exports in their inputs
will then have their inputs re-evaluated.

For more detailed information on components, see the package-level
comment in the `component` package.

Specifying Components
---------------------

Users specify a component through an HCL block. For a hypothetical
component called `remote.http`:

    remote "http" "get-url" {
        url     = "http://example.com/some/path"
        refresh = "10s"
    }

The fully-qualified identifier of this component is the combination of
the HCL block name and all of its labels: `remote.http.get-url`. This
identifier may be referenced by other components to either access any of
the input fields or any of the exported fields.

Assuming `remmote.http` exports a "response_body" field:

    remote "http" "get-second-url" {
       // http://example.com/some/path's response body contained a URL.
       url = remote.http.get-url.response_body
    }

The final label in both of these examples, `get-url` and
`get-second-url`, is called the "user identifier." It allows the user to
specify more than one instance of `remote.http` components to run.

Components may register themselves as "singletons," which means that
they do *not* support the user identifier label. Not supporting the user
identifier label means that there can only be one declaration of that
component across the entire config file.

Assuming that `node_exporter` is a singleton component:

    // No user identifier supported. The fully-qualified identifier is
    // just "node_exporter".
    node_exporter {
      ...
    }

    // Invalid: two components with the fully-qualified identifier
    // "node_exporter" are defined.
    node_exporter {
      ...
    }

Controller implementation
-------------------------

The Flow controller is implemented across a few parts:

1. `ReadFile`: a function which parses a configuration file. The
   component blocks are not interpreted at this point.

2. `buildGraph`: a function which takes the output of ReadFile and
   constructs a DAG from it, creating nodes and edges.

3. `userComponent`: the node in the Flow DAG which wraps around a
   component implementation. Each `userComponent` is responsible for
   evaluating their own HCL block and mapping it to the component
   implementation's config format.

4. `Controller`: loads and evaluates component DAGs over the process
   lifetime. Passes components to run to the scheduler. The Controller
   reacts to components who update their exports and will re-evaluate
   affected DAG nodes.

5. `scheduler`: manages goroutines for components which are expected to
   be running.

HCL evaluation
--------------

HCL does not evaluate direcly to Go. Rather, it evaluates to `go-cty`,
duck-typed system which avoids the need for Go reflection.

An `*hcl.EvalContext` is required to resolve variable or function
references in an HCL expression. EvalContexts can create a tree, where
each EvalContext may have a parent scoping for more varaibles and
functions.

We manage a `rootEvalContext` which contains only global functions to
expose to end users. Most of these functions come from `go-cty`'s
standard library, though this initial commit includes an `env` function
to get an environment variable value.

Because component implementations work with Go types, we need to map to
and from `go-cty` values to represent the current state of a component.
The `graphContext` maintains a cache of all `go-cty` values across all
components. `graphContext` can then build its current state into an
*hcl.EvalContext for other components to use.

go-cty capsule values
---------------------

`go-cty` values normally map to plain-old-data Go types, like bool or
int. It is possible to pass around arbitrary pointers in `go-cty` (and
by extension, HCL) through what it calls "capsule types" and "capsule
values." (Capsule comes from "encapsulation.")

This commit introduces an abstraction over capsule types where arbitrary
go structs can be registered as capsule types, which can then be
referenced normally like any other value.

This allows for component implementations which want to pass around a Go
interface, channel, or other values for things like pipelines.

See `component.RegisterGoStruct` for more information.

Future work
-----------

The controller is far from done. There are a few omissions in particular
which will be done in later commits:

* Because the HCL configuration file specifies LogLevel/LogFormat, we
  need the controller to manage its own logger rather than having one
  passed to it.

* The root block in the HCL configuration file is missing TLS settings
  for its server, which will be done later.

* The Controller currently doesn't detect reference cycles.

* There are no component implementations other than the testcomponents.

* There is no support for config partials (i.e., combining multiple HCL
  files into one giant Controller).

* A general strategy is needed to consider how unhealthy components
  should impact the rest of a graph. Should a node continue referencing
  the last healthy config/exports of a component that has gone
  unhealthy? Should that node become unhealthy itself? Should that node
  reference the zero value for the config/exports instead?

  This item will require the most thought, and will likely come much
  later as we built out the system.

However, I think this commit will be good for an initial implementation
that can be expanded on.

Closes #1677.