Rewrite all front matter as YAML with current alias defined
The addition of the current alias has no effect on the website as Hugo
will not create an HTML redirect if there is already content in the
web server directory.

The reason for the addition of the alias is to reduce the likelihood
of 404s occuring when files are moved.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>