remove deprecated server YAML fields in favor of static flags (#1816)
* remove deprecated server YAML fields in favor of static command line flags

Many `server` YAML fields were deprecated in v0.24.0 and scheduled for
removal in v0.26.0. This commit removes those fields in favor of just
using the flags.

This also removes the logic for determining listen address by examining
both the configured listen address and listen host/port combination, as
the latter was only available in the flags. This simplifies the fix
introduced by #1762, which can reliably look up the port used for the
gRPC server.

* remove final references to removed deprecated flags in documentation and code