pkg/river: introduce scanner package (#1852)
* pkg/river: introduce scanner package

This commit introduces the lexer for River (grafana/agent#1839). It is
not wired in anywhere, and is the first of several PRs to slowly bring
over functionality from the prototype. Most of the scanner code has been
copied from go/scanner, adapted slightly for River and decorated with
extra comments.

Note that this commit does not indicate we will ship River, just that we
are continuing the prototyping process in the main branch. pkg/river may
still be removed in the future in favor of an alternative solution.

* pkg/river: fix lint errors

* pkg/river: hush unparam warnings for Scanner.switch2

Assuming `=` as the next rune seems harder to follow than explicitly
passing it, even if it's always the same argument.

* address review feedback

* river/scanner: fix grammar for float

Float should allow exponents on numbers without a `.`

* river/scanner: denote that sign after exponent is optional