pkg/river: introduce parser package
This commit introduces the parser for River (grafana/agent#1839). It is
not wired in anywhere, and is the second of several PRs to slowly bring
over functionality from the prototype. The parser code has been highly
inspired by go/parser, with tests copied from the go/parser tests.

Note that this commit does not indicate we will ship river, just that we
are continuing the prototyping process in the main branch. pkg/river may
still be removed in the future in favor of an alternative solution for
Flow.