river: add internal/value package
This commit introduces the first pass at a value package for River
(grafana/agent#1839). It is not wired in anywhere yet, and further PRs
will improve the package and eventually wire it in.

Note that this commit does not indicate we will ship River, just that we
are continuing the prototyping process in the main branch. pkg/river may
still be removed in the future in favor of an alternative solution for
Flow.

This commit purposefully leaves three larger tasks incomplete to be done
in follow up commits:

1. Ability to invoke function values
2. Proper encoding/decoding to Go structs with rvr block tags (currently,
   labels are ignored)
3. Decoding to Go structs with missing required attributes should fail